chore: Move trackedEntityAttributeValue service and store into tracker [DHIS2-17714]
