+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/trackedentityattributevalue/TrackedEntityAttributeValueDeletionHandler.java
@@ -27,11 +27,10 @@
  */
 package org.hisp.dhis.trackedentityattributevalue;
 
-import static org.hisp.dhis.system.deletion.DeletionVeto.ACCEPT;
-
+import java.util.Map;
 import lombok.AllArgsConstructor;
-import org.hisp.dhis.system.deletion.DeletionHandler;
 import org.hisp.dhis.system.deletion.DeletionVeto;
+import org.hisp.dhis.system.deletion.JdbcDeletionHandler;
 import org.hisp.dhis.trackedentity.TrackedEntityAttribute;
 import org.springframework.stereotype.Component;
 
@@ -40,12 +39,10 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @AllArgsConstructor
-public class TrackedEntityAttributeValueDeletionHandler extends DeletionHandler {
+public class TrackedEntityAttributeValueDeletionHandler extends JdbcDeletionHandler {
   private static final DeletionVeto VETO =
       new DeletionVeto(
           TrackedEntityAttributeValue.class, "Some values are still assigned to this attribute");
-
-  private final TrackedEntityAttributeValueService attributeValueService;
 
   @Override
   protected void register() {
@@ -53,8 +50,8 @@ public class TrackedEntityAttributeValueDeletionHandler extends DeletionHandler 
   }
 
   private DeletionVeto allowDeleteTrackedEntityAttribute(TrackedEntityAttribute attribute) {
-    return attributeValueService.getCountOfAssignedTrackedEntityAttributeValues(attribute) == 0
-        ? ACCEPT
-        : VETO;
+    String sql =
+        "select 1 from trackedentityattributevalue where trackedentityattributeid = :id limit 1";
+    return vetoIfExists(VETO, sql, Map.of("id", attribute.getId()));
   }
 }
